Data Science Salary/deployment/requirements.txt
streamlit 
pandas  
seaborn 
matplotlib
Pillow
numpy
WordCloud
ImageColorGenerator
STOPWORDS
scikit-learn==1.2.2